// mobile/src/types/navigation.ts
export type Screen =
  | 'splash'
  | 'onboarding'
  | 'language'
  | 'login'
  | 'register'
  | 'country'
  | 'city'
  | 'home'
  | 'explore'
  | 'transport'
  | 'restaurants'
  | 'guide'
  | 'safety'
  | 'community'
  | 'profile'
  | 'ocr'
  | 'price'
  | 'taxiSimulator'
  | 'applyForm'
  | 'guideRequest'
  | 'activityDetail'
  | 'restaurantDetail'
  | 'checkout'
  | 'groupActivityDetail'
  | 'driverJoin'
  | 'driverPending'
  | 'driverVerify'
  | 'languageSelect'
  | 'adminDashboard'
  | 'driverDashboard'
  | 'guideDashboard'
  | 'driverLogin'
  | 'driverMessages'
  | 'driverReviews';